Manage personnel ToolBar and PopOverMenu (#35)
* half implemented cell select for edit-pane.

* added checkboxes to AddPersonnelSideSheet.
Now possible to delete manually created lines in AddPersonnelSidesheet. The general UX wil change on this.
Added delete personnel functionality to Manage Personnel Page.
Deleting users does not update the view, still a work in progress.

* updated personnel reducer with new verb delete.
Takes an array of personnel that are to be removed from local Storage.
implemented new delete verb in manage personnel page when deleting personnel.
Page now updates properly.
Also resets selectedItems when deleting.

* ToolBar now hides buttons completly if no info is passed in.
Buttons on the toolbar also takes in a color, it needs to be adjusted.
Added new menu column to AddPersonnelSideSheet
This is a popovermenus of sort.  It is a little bit rought for now.
For now it only has delete button.
Delete button is only active for lines that is not saved / previously created.

* some minor adjustmenst and clean up

* saved a bunch of files to active prettierformatting.

* reworked the popOverMenu.
Using usePopoverRef instead. This still has some issues when clicking outside the menubox
Did another pass on the edit table css. It works alot better now.
removed old popoverMenu