Api - Support for contractor change requests (#43)
Change requests will update an existing position.

To create the change requests the POST must include originalPositionId that points to an existing org chart position.

The api will validate that the position actually exists when verifying the input, and validate that the position belongs to the correct contract and project when executing the command.

The api will resolve the position targeted when returning the request object. New properties are

{
  "originalPositionId": "nullable guid",
  "originalPosition": { "same as position property" },
  "originalPerson": { "api person" }
}
When fetching all requests, the originalPosition and originalPerson properties are null. This is due to the fact we need to resolve these from the org chart to get the info. To include these properties one can call the endpoint with ?$expand=originalPosition.

Fixes: AB#10553, AB#10554