Added Scroll to top fab for add manage personnel form.
Created new component ScrollUpFab that is a button with an up arrow that show up in the bottom right corner.

Created new useScrolltoTop hook.
The hooks supplies a ref you can attach to the container you want to scroll the scroll on.
Also supplies hasScrolled, a boolean that tells you if the elemement is scrolled down based on the hasScrolledActiveLimit.
The limit is per default 100 but can be supplied as a parameter.
In this implementation its used to determin if the scroll up button should show or not.
Last it supplies a scrolltoTop function. Calling this will scroll you to the top of your container.