Feature/excel import (#140)
* bumped version number

* Api endpoint setup for  Excel import parser

* Excel Icon. Its not aligne properly. Need to fix that later

* Model for response from ExcelParser

* Added excel import button to Managepersonnel toolbar

* started building up the excel reponse mapping. This is in a pretty broken state for as off now.

* Createad excelImport hook.
Takes in an object with columns and what to search for.
Tries to map given excel file according to supplied columns.
Hook implemented in manageperonnel.

UI for import still missing

* Create Excel Import Modal.
opens when clicking excel button. Allows upload of xslx files via drag and drop or selecting file manually.
When file is selected user can press prossesing button.
Interface will show spinner until excel file is done prossesed then opens up directly into sidesheet.

* Added a reset to the modal. So that it does not open data from previous upload.
Added callback hooks to most functions in the modal.
Added autoGenerateColumns feature to excelImport.
takes in a column name and function.
Will add column to generated list regardless if its in the excel file or not.
In this case it ensure a unique uuid for each person.

* Added a excelImport flag to be sendt into the AddPersonnelSideSheet.
For now used to get around the isFormDirty flag. As the form would be marked as not dity on import.
And user was unable to save.
ExcelImport is now always marked as dirty.

* changed name off ExcelImportModal close parameter to setIsOpen.
a clearer and more consistent with app as a whole.

* Missed one close to setIsOpen change

* corrected all uses of the made up word prossessing to process or processing.

* corrected capitalization of AutoGenerateColumn

* ApiClient - ExcelImportParserAsync renamed to parseExcelFileAsync.
resouceCollection excelImportParser renamed parseExcelFile

* Moved excel Import settings for manage personnel into its own file.

* renamed getExcelReponseAsync to processExcelFileAsync.
processExcelFileAsync now contains the entire process and state handling for the selected file.
Making the supporting functions more "pure".
Also added a try/catch for the file processing.

* ExcelImportModal changed to ExcelImportSideSheet.
General look an functionality is the same.

* Updated ExcelIconImport with proper svg.

* useExcelImport error renamed for clarity.
ExcelImportSideSheet now uses the processingError and write out a generic error message if its true.
ExcelImportSideSheet css has been reworked.
It had a lot of remnants from the modal version.

* Moved table header and body creation into own memoized function.
this does not solve the performance issue, but it removed alot off rerenders.

* Rearranged some code in useExcelImport.  Just so its abit easier to read.
rewrote formatExcelReponse for loops into reduce.
rewrote mapHeaderIndexesToColumns for loop into reduce. Also moved the headers.find statement into its own functions. Which then just return the index number we are looking for.

* ExcelImportSideSheet changed ondragEnter/Over event to call function directly.

* The excelResponse in useExcelImport, might return empty lines.
This looks to happen when you have writen in a column in the file, but later removed it.
As a fix for this, data.map in formatExcelRespone has been changed to a reduce.
So I could implemented a rowIsEmpty check.
That then removes the line if its empty.

* rewrite of the PersonnelForm. WORK IN PROGRESS.
This fixes the problem with big delays when typing, when working with a large dataset.
It still have problem with scrolling.

* Moved parent formstate handling up a level.

* Added save in progress to saveinprogresds as dependency to renderFormBody.
Removed a console log from addPersonnelSideSheet

* removed select all State. As it was not needed.
changed formState.length dependencty on renderFormBody to be only formState.
the length part caused problem with clojures, so that you only got changes from one person saved back to the formstate.
This changed did not impact performance for typing.

* Moved save button into own component for readablity

* reverting formState.lenght change.

* Update components with fix for scrolling in Add Personnel Form.
New handling for the selections fields forcing a full refresh when selecting/deselecting all from header.

* added in som missing dependecies.
Renamed AddPersonnelFormLine to AddPersonnelFormRow.

* basePosition was not updated in the dropdown first time you tried to open a Add Person form.
Added dependency on basepositions to renderFormBody to fix this.

* Added Scroll to top fab for add manage personnel form.

Created new component ScrollUpFab that is a button with an up arrow that show up in the bottom right corner.

Created new useScrolltoTop hook.
The hooks supplies a ref you can attach to the container you want to scroll the scroll on.
Also supplies hasScrolled, a boolean that tells you if the elemement is scrolled down based on the hasScrolledActiveLimit.
The limit is per default 100 but can be supplied as a parameter.
In this implementation its used to determin if the scroll up button should show or not.
Last it supplies a scrolltoTop function. Calling this will scroll you to the top of your container.

* Row number column added to add personnel form.

* excel import. Added excel template download.
New apiClient endpoint fetching the excel template.
Added a button in sidesheet that fetches and downloads the excel template.
Added a new Tips section to the sidesheet, with the button to download the template. And som pointer for the users.

* changed help to tips for consistensy

* Added Batch import of personnel.
Used when importing more than 10 personnel at the same time.

* changed batched to send batches of 100 personnel per request.

* Moved ReadableStreamReponse model into its own file

* scollUpFab styling. Changed size to use grid-unit instead off direct px

* Removed extra semi-colons from.