Api - Support for resolving task owner from org chart (#234)
Updated mock implementation to do actual resolving of the task owner from the org chart. 

Added error handling to the expand, so request does not fail completely. Consumer should monitor if the property is still null even after expanded.
When the task owner is resolved the api model will contain the date used when calculating, the position id located, a list of instance ids that are active and a list of persons assigned to the respective instances.

```json
{
   "date": "DATETIMEZ",
   "positionId": "Guid?",
   "instanceIds": [ "GUID" ],
   "persons": [ { } ]
}
```

The collection for instances and persons are required as the task owner could be a rotation position. 

The date used to calculate the task owner depends on the instance dates. Current date if > instance.AppliesFrom and < instance.AppliesTo. If out of bounds on any side, the instance from/to date is used.