Updated department personnel query and allocation metadata (#255)
As the search index now has the manager property we can query on this property to get the personnel reporting to a department. This will give us a better view as the child department managers have their own department on the profile.
Added display of the new allocation state properties, which should indicate to the resource owner that the "task" has been updated after their person was allocated.