Reset allocation state endpoint (#258)
Added endpoint to reset allocation state on a persons allocation. This should be used by the resource owner to "accept" changes done by the task owner.
The operation will fetch the allocations for the user by using the profile search. From here the project and position id is resolved. Only the person id and instance id must be provided by the client.