Person notes (#263)
Added endpoints to support adding notes to a resources. Instead of just one note the solution is based on a collection of notes, which can be labeled and support for marking specific items as shared. 
The idea is that some notes could be of a private nature, while others could be shared with other resource owners in the area (ex. what the person wants to work with etc). The public info could then be used in a search engine when locating candidates.

Endpoints are:
- GET|POST `/persons/{uniqueId|mail}/resources/notes`
- PUT|DELETE `/persons/{uniqueId|mail}/resources/notes/{noteId}`

Supported api version: `1.0-preview`