Person notes authorization (#274)
Added authroization to endpoints related to person notes. 
These are basic rules related to the resource owner flag and how they are in relations to the person the note exists at. 

The direct resource owner is allowed to manage the notes. 
Shared notes are visible for resource owners within the same L2. Ex. `L1 A DEP1` and `L1 A DEP2` is allowed, but `L1 B *` is not allowed.