Updated to instance based task owner endpoint (#283)
In some cases the task owner is incorrectly returned from resources. This might be based on manually calulating the relevant date. There is an endpoint that is specifically made to get the relevant taskowner for an exact instance, using either start/end or todays date.
Change expand task owner functionality in specific request query, to use the instance based endpoint.