Disable change requests for personnel in projects not in pilot (#301)
As some projects in pilot will be synced TO pims, while the rest is synced FROM pims, we must disable the possibility for resource owners to create change requests on allocations that will be overwritten from pims nightly.
The projects with pims write sync enabled should be safe to create change requests on.
We can do this by checking the property on the project. This is however a temporary solution in the pilot phase.

A check has been placed on `/projects/:orgProject/positions/:orgPositionId/instances/:orgInstanceId/resources/requests?requestType=resourceOwnerChange`. This will return `POST` in the `Allowed` header if request type can be created. For now only the `resourceOwnerChange` type is supported. If the request type is not supported, a bad request with `InvalidInput` will be returned. The endpoint can return a error payload with `error.code` and `error.message` to give further details if no allowed headers.