Disabled aks deploy for prod api (#352)
After update to latest kubernetes version the ingres version had to be updated to latest non-beta version. This has only been done on the test cluster, not the production cluster.
This makes the deploy to production cluster step fail. As this is only a pilot and not used by production, this should be disabled untill the production cluster has been recreated.