Pipeline for rotating client secret (#410)
Added scripts to manage client secrets used by the app. The script will generate client secret and store it in key vault as a secret. The KV secret will be taged and marked with expiration date.
Another script will loop through KV secret versions and try to remove the credential if expired. The credential to remove is located by using the keyId tag added to the KV secret, which is the id for the credential on the app reg.

Updated pipelines that reference old secret in shared key vault, to use env specific key.