Removed dependency to shared client secret (#411)
The initial solution was set up to depend on a shared secret as a quick fix. However the permanent solution was never implemented.
As a new secret pr env has been set up using a rotation pipeline, the shared secret should no long be used, as it is not renewed.