Added new query to resolve persons that belong to a specific department (#571)
Problems have been identified with the query used to resolve employees in a department, when changing to a context based personnel app.
We cannot use the manager as reference as there are no guarantee that the manager in only manager for the specific department. A person can be manager for multiple departments at the same time, in cases where one has not yet been employeed.
When there is no maanger, the responsibility falls to the next level up, in SAP, which is synced to AD.

The v1 view is working with the current app where the manager will get all departments in the same view.. If we update the v1 endpoint, the current app will not work. We cannot publish the new app, as this would make it unusable for the users with multiple departments.

Solution is to introduce a v2 endpoint, which the new app will utilize. Then when old app is no longer, we can change how v1 endpoints return data.