Fixed template to add web app config for portal url in production (#573)
When users receive notifications which lets them open relevant resources in the portal, the hostname is using the old value which is configured to redirect users. 
This creates issues when the browser has to run an signin session at the same time, due to oauth flow.

The portal url is resolved through integration lib. This should be updated to the latest version, but this requires more testing. 

Added quick fix, where resolved host can be overridden through web app config. Updated the web app template for prod to set fusion.equinor.com as hostname for portal.