feat(api): Ensure data is committed to database before public events are dispatched (#572)
Noticed events are dispatched on event bus before data is commited to database. This means external parties can react and query data before this is available.

Fixed by utilizing the transaction feature available in fusion event library. This will store events internally on a queue in memory and only dispatch on the event bus when the transaction is commited.
~~Created new transaction object that wraps both db and event transaction.~~ Since the transaction feature in event lib uses AsyncLocal the scope has to be created in the same async method as the rest of the code executes. If this is done in a  'branch' the AsyncLocal scope will be cleared when continuing to the next statement in the method.