Remove positions managed by contract personnel (#585)
To support legacy functionality and downstream apis, the contract personnel solution is generating "virtual" positions for users which is connected to contract roles. These positions does not exist in the org charts but are caused by formal roles related to contracts.

The problem is related to managers trying to remove thsese roles through the personnel allocation app. This is not possible as the ids do not exist.

The fix consist of filtering away these positions when we fetch data from the search index.