feat: Remove external personnel from summary notification (#619)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
Have removed external personnel (users with accountType=consultant) from
summary notifications so that these are not included in the KPI's.
Did also small adjustments regarding getting leave for calculating
capacity. It was not necessarry to get this from an own endpoint so I
removed it. The calculation is the same.


[AB#49281](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/49281)



**Testing:**
- [x] Can be tested
- [x] Automatic tests created / updated
- [x] Local tests are passing

Can be tested with running summary notification and verifying that all
personnel of type "consultant" has been removed from the notification.
It should be possible to find which users are external personnel by
checking in Personnel Allocation for a specific department. All of the
external personnel should have the name of their affiliated company in
parentesies (example: FirstName Lastname (Bouvet ASA)



**Checklist:**
- [x] Considered automated tests
- [ ] ~~Considered updating specification / documentation~~
- [x] Considered work items 
- [ ] ~~Considered security~~
- [x] Performed developer testing
- [x] Checklist finalized / ready for review