feat/Add cache of positionIds not found when getting requests (timeline) (#627)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
When getting requests via
/departments/{departmentString}/resources/requests/timeline we try to
resolve all referenced Positions by checking the OrgPositionId. Because
the request may reference deleted positions this may result in 404 in
the Org-API. Some endpoints in Resource-API is returning alot of
requests and this may also result in many 404s in Org-API. To avoid this
we cache all OrgPositionId which returns 404 and don't check them again.

The timeout for cache has been set to 7 days.

WorkItem:

[AB#44789](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/44789)

Related pr: 
https://github.com/equinor/fusion-app-resources/pull/624

**Testing:**
- [x] Can be tested
- [ ] Automatic tests created / updated
- [x] Local tests are passing

Not that easy to test as this is to reduce the errors in application
insight. But we should still get the same data when trying to get all
the requests. It should only reduce the errors in application insight.

Use endpoint in postman to get all request within a timeline:
GET:
https://resources-api-pr-627.fusion-dev.net/departments/PDP%20PRD%20PMC%20PCA%20PCA1/resources/requests/timeline?timelineStart=2024-04-01T00%3A00%3A00&timelineEnd=2024-09-30T00%3A00%3A00&api-version=1.0)



**Checklist:**
- [ ] Considered automated tests
- [ ] Considered updating specification / documentation
- [x] Considered work items 
- [ ] ~~Considered security~~
- [x] Performed developer testing
- [x] Checklist finalized / ready for review