feat: Adjusted the calculation of the field "Project changes last week affecting next 3 months" for a department to use AllocationState/AllocationUpdated (#628)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
Because of the big load time when getting from /change-log from Org-API
we now use the fields AllocationState and AllocationUpdated to calculate
which instanses that have been updated the last week. We also only
include active instances


[AB#51384](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/51384)

**Testing:**
- [x] Can be tested
- [x] Automatic tests created / updated
- [x] Local tests are passing
We need to execute the function
"scheduled-report-timer-trigger-function" and then we need to verify
that the field "Project changes last week affecting next 3 months" is
correct. This is done by checking a the notifications for departments
and verify that the number of changes matches the field "Project changes
last week affecting next 3 months"

We also need to check the execution time

**Checklist:**
- [x] Considered automated tests
- [x] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review