fix (resource): Adjust AverageTimeToHandleRequests field on summary notification  (#631)
- [ ] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**
The field "Average time to handle request" on the Summary Notification
did not give a correct number. The reason was that we checked for that
the requests was in state == completed, meaning that we did not include
requests that where finished from all parts (resourceowners and
taskowners). We want to include requests that have been handled by
ResourceOwner, but also those that are still open.
We have also adjusted so that we include requests for the last 12 months
(instead of 3 months).
If we don't find the date for proposal we will use todays date for when
calculating the amount of time used to handle a request


[AB#52911](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/52911)

**Testing:**
- [x] Can be tested
- [ ] Automatic tests created / updated
- [x] Local tests are passing

Can be tested with running Summary Notifications and check that the
value on "Average time to handle request" is correct. May need to
manually calculate this value for verify this. See the work item for
description on how to test this.

**Checklist:**
- [ ] Considered automated tests
- [ ] Considered updating specification / documentation
- [ ] ~~Considered work items~~ 
- [ ] ~~Considered security~~
- [x] Performed developer testing
- [x] Checklist finalized / ready for review