fix: Reverted to old way of finding creation date and approval date (#632)
- [ ] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**
Needed to revert to old way of finding the dates of the different
states. We need the Workflow.Steps to find the exact date for when a
request moves from "Created" to "Proposed", and not just use the date
for when a person is proposed (this can be done by the task owner also)


[AB#52911](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/52911)

**Testing:**
- [x] Can be tested
- [ ] Automatic tests created / updated
- [x] Local tests are passing

Can be tested with running Summary Notifications and check that the
value on "Average time to handle request" is correct. May need to
manually calculate this value for verify this. See the work item for
description on how to test this.

**Checklist:**
- [ ] Considered automated tests
- [ ] Considered updating specification / documentation
- [ ] ~~Considered work items~~ 
- [ ] ~~Considered security~~
- [x] Performed developer testing
- [x] Checklist finalized / ready for review