feat: establishing summary function resource (#633)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Work description:**

This work establishes a new resource from which we can add more logic to
in the future. Fusion.Summary.Functions will be the home for all
functions related to Summary (Summary Notification).

_Expected result_: A new resource is established in the PR environment

Link to workitem:
[#53068](https://statoil-proview.visualstudio.com/Fusion%20Resource%20Allocation/_workitems/edit/53068)
 
**Test description:**
- [x] Can be tested
- [ ] Automatic tests created / updated
- [ ] Local tests are passing

Before finalizing there should be two functions. A anonymous http get
endpoint that outputs a welcome message and a timer (10s) that should be
seen in the logs.
 
**Checklist:**
- [x] Considered automated tests
- [ ] Considered updating specification / documentation
- [x] Considered work items 
- [ ] Considered security
- [x] Performed developer testing
- [ ] Checklist finalized / ready for review

---------

Co-authored-by: Hans Dahle <handah@equinor.com>
Co-authored-by: aleklundeq <138589813+aleklundeq@users.noreply.github.com>
Co-authored-by: Aleksander Lund <alun@equinor.com>