feat: Update assigned department on request when org unit is renamed (#639)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
When org units are renamed or moved we still persist the old full
department string. This can leave requests connected to org units that
no longer exists -> orphan requests.

We can subscribe for events comming from line org to act when these
changes are done. This should be done as a persistant event handler so
we are guaranteed delivery in case of downtime etc.

- Added persistant event handler to perform sync on assigned department
for requests

The transient handler are used to invalidate cache, as this is executed
on all instances of the api.


**Testing:**
- [x] Can be tested
- [ ] ~~Automatic tests created / updated~~
- [x] Local tests are passing

Can be verified by assigning a ticket to some department (use one not
commonly used, so we dont cause much problems).
PATCH the org unit in lineorg, chaning the full department (add a
character etc).
Verify requests are assigned to that department
PATCH the org unit back to the original value (clean up)
Can verify requests are assigned back


**Checklist:**
- [ ] ~~Considered automated tests~~
- [ ] ~~Considered updating specification / documentation~~
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review


[AB#53786](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/53786)
Limited changes to only requests. There are other entities which stores
full department reference. These should be changed in own prs.