feat: Clean up requests when org unit is deleted (#640)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
As we get more requests and the organisation is re-organising we gather
more and more data which is connected to deleted references, e.g deleted
org units.

We should try and delete data that will never be relevant (draft
requests) and "release" requests that are connected to non-existing
units. That way they are at least known to be unassigned.

Updated persistant event handler to publish an `OrgUnitDeleted`
notification and two handlers which do solve seperate functionality.
Wanted to divide this into smaller functions to increase readability.

Updated "editor" concept to support SystemAccount, when HttpContext is
not present. Using Guid.Empty here so we can easily catch it and process
it seperately as it will not resolve to an active profile.


**Testing:**
- [x] Can be tested
- [ ] ~~Automatic tests created / updated~~
- [x] Local tests are passing

To verify, one can create a new org unit in line org, with a new sap id.
Create different requests to verify functionality. The allocation
requests can be "reassigned" using the UI, while the change request must
be created in a
different org unit which has employee, then updated using PATCH
/.../:requestNumner.
Then the org unit can be deleted in line org and changes verified /
confirmed not affected.

- change request which has been sent - Should not be changed at all
- draft change request - should be deleted
- allocation request - should be unassigned


**Checklist:**
- [ ] ~~Considered automated tests~~
- [ ] ~~Considered updating specification / documentation~~
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review


[AB#53894](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/53894)