feat: Summary AZ Function code setup and DepartmentResourceOwnerSync (#645)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Work description**:

Created a class library that contains common logic for the two function
apps. It's namespace is `Fusion.Resources.Functions.Common`, I feel like
there is a better name for it, but cant think of one.

Added DepartmentResourceOwnerSync timer function that retrieves
DepartmentResourceOwners for LineOrg and sends them to the new Summary
API. It does not put anything on the buss as of now, thinking of adding
that in a separate PR.

There's also the question of wheter the summary app should have it's own
application registration. As of now it uses resources' registration. But
this result in minor code changes for this PR.

**The `department-resource-owner-sync` is disabled because the Summary
API still isn't merged in / being worked on**

Link to workitem: US
[AB#51132](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/51132),
task
[AB#54679](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/54679)

**Test description**:

- [ ] Can be tested
- [ ] Automatic tests created / updated
- [x] Local tests are passing


**Checklist**:

- [x] Considered automated tests
- [ ] Considered updating specification / documentation
- [x] Considered work items
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review