feat: Auto approve direct allocations that have no proposed changes (#659)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->

[AB#50816](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/50816)

For direct allocation requests, if there are no proposed changes from
the Resource Manager, then it will be auto approved by the project.


- When a allocation request is created the initially proposed person is
saved to the db. This is needed as earlier we did not save the first
person.

The UI text is subject to change.

**Testing:**
- [x] Can be tested
- [x] Automatic tests created / updated
- [x] Local tests are passing

<!--- Please give a description of how this can be tested --->

Had to modify some of the existing allocation requests. Created some new
ones as well


**Checklist:**
- [x] Considered automated tests
- [x] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review

<!--- Other comments --->