fix: Summary report email layout (#676)
- [ ] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**


[AB#54653](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/54653)

This PR rewrites the adaptive card code to make it look better. It also
adds personnel-allocation app key to the notifcation. This means that if
core adds a toggle for personnel allocation, then emails can be turned
of for the app. https://github.com/equinor/fusion/issues/390

If we want to perfect the mail design then I believe we'll need to
create a separate message and manually send it to the mail service. As I
couldn't get the list values to show properly and I couldn't find a way
to center the action button. See US
https://statoil-proview.visualstudio.com/Fusion%20Resource%20Allocation/_workitems/edit/55681

New design:

![image](https://github.com/user-attachments/assets/6ecb4297-f205-4073-a764-20386bcaef97)


Old design:

![image](https://github.com/user-attachments/assets/2e664e68-a524-4d8d-877e-baa1fd280345)


**Testing:**
- [x] Can be tested
- [ ] Automatic tests created / updated
- [x] Local tests are passing



**Checklist:**
- [x] Considered automated tests
- [ ] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review